Move scraper to data_access layer
