--- newspeek/src/test/java/use_case/helpers/JReadabilityScraperTest.java
@@ -1,7 +1,0 @@
-package use_case.helpers;
-
-import org.junit.jupiter.api.Test;
-
-class JReadabilityScraperTest {
-
-}
